@@ -1,5 +1,4 @@
 import configparser
-import io
 import logging
 import os
 import re
@@ -15,7 +14,7 @@ from cookietemple.util.rich import console
 log = logging.getLogger(__name__)
 
 
-class TemplateLinter(object):
+class TemplateLinter:
     """Object to hold linting information and results.
     Attributes:
         files (list): A list of files found during the linting process.
@@ -153,19 +152,21 @@ class TemplateLinter(object):
 
     def check_docker(self):
         """
-        Checks that Dockerfile contains the string ``FROM``
+        Checks that Dockerfile exists and contains the string ``FROM``
         """
         fn = os.path.join(self.path, "Dockerfile")
-        with open(fn, "r", encoding="utf-8") as fh:
-            content = fh.read()
-
-        # Implicitly also checks if empty.
-        if "FROM " in content:
-            self.passed.append(("general-2", "Dockerfile check passed"))
-            self.dockerfile = [line.strip() for line in content.splitlines()]
-            return
-
-        self.failed.append((2, "Dockerfile check failed"))
+        if os.path.isfile(fn):
+            with open(fn, encoding="utf-8") as fh:
+                content = fh.read()
+                # Implicitly also checks if empty.
+                if "FROM" in content:
+                    self.passed.append(("general-2", "Dockerfile check passed"))
+                    self.dockerfile = [line.strip() for line in content.splitlines()]
+                    return
+                else:
+                    self.failed.append(("general-2", "Dockerfile check failed"))
+        else:
+            self.failed.append(("general-2", "Dockerfile check failed"))
 
     def check_cookietemple_todos(self) -> None:
         """
@@ -173,7 +174,7 @@ class TemplateLinter(object):
         """
         ignore = [".git"]
         if os.path.isfile(os.path.join(self.path, ".gitignore")):
-            with io.open(os.path.join(self.path, ".gitignore"), "rt", encoding="latin1") as file:
+            with open(os.path.join(self.path, ".gitignore"), encoding="latin1") as file:
                 for line in file:
                     ignore.append(os.path.basename(line.strip().rstrip("/")))
         for root, dirs, files in os.walk(self.path):
@@ -184,7 +185,7 @@ class TemplateLinter(object):
                 if ignore_file in files:
                     files.remove(ignore_file)
             for fname in files:
-                with io.open(os.path.join(root, fname), "rt", encoding="latin1") as file:
+                with open(os.path.join(root, fname), encoding="latin1") as file:
                     for line in file:
                         if any(todostring in line for todostring in ["TODO COOKIETEMPLE:", "COOKIETEMPLE TODO:"]):
                             line = (
@@ -208,7 +209,7 @@ class TemplateLinter(object):
         """
         for root, _dirs, files in os.walk(self.path):
             for fname in files:
-                with io.open(os.path.join(root, fname), "rt", encoding="latin1") as file:
+                with open(os.path.join(root, fname), encoding="latin1") as file:
                     if file.name.endswith(".pyc"):
                         continue
                     for line in file:
@@ -347,7 +348,7 @@ def files_exist_linting(
     for files in files_fail:
         if not any([os.path.isfile(pf(self, f)) for f in files]):
             all_exists = False
-            self.failed.append(("{handle}-1", f"File not found: {self._wrap_quotes(files)}"))
+            self.failed.append((f"{handle}-1", f"File not found: {self._wrap_quotes(files)}"))
     # flag that indiactes whether all required files exist or not
     if all_exists:
         # called linting from a specific template linter
