@@ -38,7 +38,7 @@ def test_bump_version(mocker, valid_version_bumpers) -> None:
     version_bumper = VersionBumper(Path(str(os.path.abspath(os.path.dirname(__file__)))), downgrade=False)
 
     for version in valid_version_bumpers:
-        version_bumper.bump_template_version(version, Path(str(os.path.abspath(os.path.dirname(__file__)))))
+        version_bumper.bump_template_version(version, Path(str(os.path.abspath(os.path.dirname(__file__)))), False)
         versions_whitelisted, versions_blacklisted = get_file_versions_after_bump(Path.cwd())
 
         assert (
@@ -81,11 +81,11 @@ def get_file_versions_after_bump(cwd: Path) -> Tuple[list, list]:
     """
     valid_version_regex = r"(?<!\.)\d+(?:\.\d+){2}(?:-SNAPSHOT)?(?!\.)"
 
-    with open(f"{cwd}/bump_version_test_files/bump_test_file_whitelisting", "r") as bumped_file_whitelisted:
+    with open(f"{cwd}/bump_version_test_files/bump_test_file_whitelisting") as bumped_file_whitelisted:
         bumped_data = bumped_file_whitelisted.read()
         bumped_versions_whitelisted = re.findall(valid_version_regex, bumped_data)
 
-    with open(f"{cwd}/bump_version_test_files/bump_test_file_blacklisting", "r") as bumped_file_blacklisted:
+    with open(f"{cwd}/bump_version_test_files/bump_test_file_blacklisting") as bumped_file_blacklisted:
         bumped_data = bumped_file_blacklisted.read()
         bumped_versions_blacklisted = re.findall(valid_version_regex, bumped_data)
 
